1.协议变更 2.存储优化 3.user系统 4.LoginMgr模块 5.添加配置加载模块 6.添加公共参数配置 7.guidmgr模块 8.guidutil优化 time减少到31位 9.导表工具支持枚举数据类型 10.machineid 11.
Former-commit-id: 0ac9381d9136f4513945857f57bc5ae342e07874

+++ service_common/application/Application.cpp
@@ -902,7 +902,7 @@ Int32 Application::_ReadBaseConfigs()
                     return Status::ConfigError;
                 }
             }
-            _appConfig._machineId = static_cast<UInt16>(cache);
+            _appConfig._machineId = static_cast<UInt32>(cache);
         }
         {// 注册成功机器时间
             UInt64 cache;
